Updating fork.

+++ starclass/taskmanager.py
@@ -11,13 +11,14 @@ import numpy as np
 import os
 import sqlite3
 import logging
+from astropy.table import Table
 
 class TaskManager(object):
 	"""
 	A TaskManager which keeps track of which targets to process.
 	"""
 
-	def __init__(self, todo_file, cleanup=False):
+	def __init__(self, todo_file, cleanup=False, overwrite=False):
 		"""
 		Initialize the TaskManager which keeps track of which targets to process.
 
@@ -25,6 +26,7 @@ class TaskManager(object):
 			todo_file (string): Path to the TODO-file.
 			cleanup (boolean): Perform cleanup/optimization of TODO-file before
 			                   during initialization. Default=False.
+			overwrite (boolean): Overwrite any previously calculated results. Default=False.
 
 		Raises:
 			IOError: If TODO-file could not be found.
@@ -38,7 +40,7 @@ class TaskManager(object):
 
 
 		# Keep a list of all the possible classifiers here:
-		self.all_classifiers = set(['rfgc'])
+		self.all_classifiers = set(['rfgc', 'slosh', 'foptics', 'xgb'])
 
 		# Setup logging:
 		formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
@@ -55,8 +57,9 @@ class TaskManager(object):
 
 		# Reset the status of everything for a new run:
 		# TODO: This should obviously be removed once we start running for real
-		self.cursor.execute("DROP TABLE IF EXISTS starclass;")
-		self.conn.commit()
+		if overwrite:
+			self.cursor.execute("DROP TABLE IF EXISTS starclass;")
+			self.conn.commit()
 
 		# Create table for diagnostics:
 		self.cursor.execute("""CREATE TABLE IF NOT EXISTS starclass (
@@ -64,7 +67,8 @@ class TaskManager(object):
 			classifier TEXT NOT NULL,
 			status INT NOT NULL,
 			class TEXT,
-			prob REAL
+			prob REAL,
+			FOREIGN KEY (priority) REFERENCES todolist(priority) ON DELETE CASCADE ON UPDATE CASCADE
 		);""")
 		self.cursor.execute("CREATE INDEX IF NOT EXISTS priority_classifier_idx ON starclass (priority, classifier);")
 		self.conn.commit()
@@ -125,10 +129,30 @@ class TaskManager(object):
 		if task:
 			task = dict(task)
 			task['classifier'] = cl
+
+			# Add things from the catalog file:
+			#catalog_file = os.path.join(????, 'catalog_sector{sector:03d}_camera{camera:d}_ccd{ccd:d}.sqlite')
+			# cursor.execute("SELECT ra,decl as dec,teff FROM catalog WHERE starid=?;", (task['starid'], ))
+			#task.update()
+
+			# If the classifier that is running is the meta-classifier,
+			# add the results from all other classifiers to the task dict:
+			if cl == 'meta':
+				self.cursor.execute("SELECT classifier,class,prob FROM starclass WHERE priority=? AND classifier != 'meta';", (task['priority'], ))
+
+				#for row in self.cursor.fetchall():
+
+
+				task['other_classifiers'] = Table(
+					rows=self.cursor.fetchall(),
+					names=('classifier', 'class', 'prob'),
+					dtype=('S256', 'S256', 'float32')
+				)
+
 			return task
 		return None
 
-	def get_task(self, classifier=None):
+	def get_task(self, classifier=None, change_classifier=True):
 		"""
 		Get next task to be processed.
 
@@ -147,7 +171,7 @@ class TaskManager(object):
 
 		# If no task is returned for the given classifier, find another
 		# classifier where tasks are available:
-		if task is None:
+		if task is None and change_classifier:
 			# Make a search on all the classifiers, and record the next
 			# task for all of them:
 			all_tasks = []
